requirements.txt
    # requirements.txt
    # Python package dependencies for the EVE Tracker backend.

    Flask>=2.0 # Core web framework
    Flask-SQLAlchemy>=2.5 # Database ORM integration
    Flask-Migrate>=3.0 # For handling database schema changes (optional but recommended)
    python-dotenv>=0.19 # For loading .env files
    requests>=2.25 # For making HTTP requests to ESI
    requests-oauthlib>=1.3 # For handling the OAuth 2.0 flow with EVE SSO
    PyJWT[crypto]>=2.0 # For decoding and verifying EVE SSO JWT tokens
    cryptography>=3.4 # For encrypting/decrypting refresh tokens securely
    gunicorn>=20.0 # WSGI HTTP Server for production deployment (optional for dev)
    Werkzeug>=2.0 # For password hashing (usually installed with Flask)

    # Add database driver if not using SQLite (choose one):
    # psycopg2-binary # For PostgreSQL
    # mysqlclient # For MySQL/MariaDB

    